requirements.txt
gradio==5.29.*
PyPDF2==3.0.*
python-docx==1.1.*
google-genai==1.13.*
dotenv==1.1.0